+++ src/InternalApi/Exceptions/InvalidMessageException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Hyvor\Internal\InternalApi\Exceptions;
+
+class InvalidMessageException extends \Exception
+{
+
+}