+++ src/InternalApi/Testing/CallsInternalApi.php
@@ -9,6 +9,9 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\App;
 use Illuminate\Testing\TestResponse;
 
+/**
+ * Use this to test the internal API implementation of the current component
+ */
 trait CallsInternalApi
 {
 
@@ -19,12 +22,10 @@ trait CallsInternalApi
      */
     public function internalApi(
         InternalApiMethod|string $method,
-        string                   $endpoint,
-        array                    $data = [],
-        ?ComponentType           $from = null,
-    ): TestResponse
-    {
-
+        string $endpoint,
+        array $data = [],
+        ?ComponentType $from = null,
+    ): TestResponse {
         assert(App::environment('testing'), 'This method can only be called in the testing environment');
 
         if (is_string($method)) {
@@ -46,7 +47,6 @@ trait CallsInternalApi
                 'HTTP_X-Internal-Api-To' => ComponentType::current()->value,
             ]
         );
-
     }
 
 }