+++ tests/Unit/Auth/AuthFakeTest.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace Hyvor\Internal\Tests\Unit\Auth\Providers;
+namespace Hyvor\Internal\Tests\Unit\Auth;
 
-use Hyvor\Internal\Auth\AuthUser;
-use Hyvor\Internal\Auth\Providers\Fake\AuthFake;
-use Hyvor\Internal\Auth\Providers\AuthProviderInterface;
+use Hyvor\Internal\Auth\Auth;
+use Hyvor\Internal\Auth\AuthFake;
 use Hyvor\Internal\Tests\TestCase;
 
 /**
@@ -14,7 +13,7 @@ class AuthFakeTest extends TestCase
 {
     private function provider(): AuthFake
     {
-        $provider = app(AuthProviderInterface::class);
+        $provider = app(Auth::class);
         assert($provider instanceof AuthFake);
         return $provider;
     }
@@ -33,6 +32,7 @@ class AuthFakeTest extends TestCase
 
     public function testDatabaseHelperFunctions(): void
     {
+        AuthFake::enable();
         AuthFake::databaseSet([
             ['id' => 1, 'name' => 'John'],
             ['id' => 2, 'name' => 'Jane'],
@@ -60,7 +60,8 @@ class AuthFakeTest extends TestCase
 
     public function testFromId(): void
     {
-        $id20 = $this->provider->fromId(20);
+        AuthFake::enable();
+        $id20 = $this->provider()->fromId(20);
         $this->assertNotNull($id20);
         $this->assertIsString($id20->name);
         $this->assertEquals(20, $id20->id);
@@ -71,18 +72,19 @@ class AuthFakeTest extends TestCase
             ['id' => 2, 'name' => 'Jane'],
         ]);
 
-        $id1 = $this->provider->fromId(1);
+        $id1 = $this->provider()->fromId(1);
         $this->assertNotNull($id1);
         $this->assertEquals('John', $id1->name);
         $this->assertEquals(1, $id1->id);
 
-        $id3 = $this->provider->fromId(3);
+        $id3 = $this->provider()->fromId(3);
         $this->assertNull($id3);
     }
 
     public function testFromEmail(): void
     {
-        $email20 = $this->provider->fromEmail('20@test.com');
+        AuthFake::enable();
+        $email20 = $this->provider()->fromEmail('20@test.com');
         $this->assertNotNull($email20);
         $this->assertIsString($email20->name);
         $this->assertEquals('20@test.com', $email20->email);
@@ -93,18 +95,19 @@ class AuthFakeTest extends TestCase
             ['id' => 2, 'name' => 'Jane', 'email' => 'jane@test.com']
         ]);
 
-        $email1 = $this->provider->fromEmail('john@test.com');
+        $email1 = $this->provider()->fromEmail('john@test.com');
         $this->assertNotNull($email1);
         $this->assertEquals('John', $email1->name);
         $this->assertEquals('john@test.com', $email1->email);
 
-        $email3 = $this->provider->fromEmail('supun@test.com');
+        $email3 = $this->provider()->fromEmail('supun@test.com');
         $this->assertNull($email3);
     }
 
     public function testFromUsername(): void
     {
-        $username20 = $this->provider->fromUsername('user20');
+        AuthFake::enable();
+        $username20 = $this->provider()->fromUsername('user20');
         $this->assertNotNull($username20);
         $this->assertIsString($username20->name);
         $this->assertEquals('user20', $username20->username);
@@ -115,18 +118,19 @@ class AuthFakeTest extends TestCase
             ['id' => 2, 'name' => 'Jane', 'username' => 'jane']
         ]);
 
-        $username1 = $this->provider->fromUsername('john');
+        $username1 = $this->provider()->fromUsername('john');
         $this->assertNotNull($username1);
         $this->assertEquals('John', $username1->name);
         $this->assertEquals('john', $username1->username);
 
-        $username3 = $this->provider->fromUsername('supun');
+        $username3 = $this->provider()->fromUsername('supun');
         $this->assertNull($username3);
     }
 
     public function testFromIds(): void
     {
-        $ids = $this->provider->fromIds([1, 2, 3]);
+        AuthFake::enable();
+        $ids = $this->provider()->fromIds([1, 2, 3]);
         $this->assertCount(3, $ids);
         $this->assertEquals(1, $ids[1]->id);
         $this->assertEquals(2, $ids[2]->id);
@@ -138,7 +142,7 @@ class AuthFakeTest extends TestCase
             ['id' => 2, 'name' => 'Jane', 'username' => 'jane']
         ]);
 
-        $ids = $this->provider->fromIds([1, 2, 3]);
+        $ids = $this->provider()->fromIds([1, 2, 3]);
         $this->assertCount(2, $ids);
         $this->assertEquals(1, $ids[1]->id);
         $this->assertEquals(2, $ids[2]->id);
@@ -146,7 +150,8 @@ class AuthFakeTest extends TestCase
 
     public function testFromUsernames(): void
     {
-        $usernames = $this->provider->fromUsernames(['user1', 'user2', 'user3']);
+        AuthFake::enable();
+        $usernames = $this->provider()->fromUsernames(['user1', 'user2', 'user3']);
         $this->assertCount(3, $usernames);
         $this->assertEquals('user1', $usernames['user1']->username);
         $this->assertEquals('user2', $usernames['user2']->username);
@@ -158,7 +163,7 @@ class AuthFakeTest extends TestCase
             ['id' => 2, 'name' => 'Jane', 'username' => 'jane']
         ]);
 
-        $usernames = $this->provider->fromUsernames(['john', 'jane', 'supun']);
+        $usernames = $this->provider()->fromUsernames(['john', 'jane', 'supun']);
         $this->assertCount(2, $usernames);
         $this->assertEquals('john', $usernames['john']->username);
         $this->assertEquals('jane', $usernames['jane']->username);
@@ -166,7 +171,8 @@ class AuthFakeTest extends TestCase
 
     public function testFromEmails(): void
     {
-        $emails = $this->provider->fromEmails(['user1@test.com', 'user2@test.com']);
+        AuthFake::enable();
+        $emails = $this->provider()->fromEmails(['user1@test.com', 'user2@test.com']);
         $this->assertCount(2, $emails);
         $this->assertEquals('user1@test.com', $emails['user1@test.com']->email);
         $this->assertEquals('user2@test.com', $emails['user2@test.com']->email);
@@ -177,7 +183,7 @@ class AuthFakeTest extends TestCase
             ['id' => 2, 'name' => 'Jane', 'email' => 'jane@test.com']
         ]);
 
-        $emails = $this->provider->fromEmails(['john@test.com', 'jane@test.com', 'roger@test.com']);
+        $emails = $this->provider()->fromEmails(['john@test.com', 'jane@test.com', 'roger@test.com']);
         $this->assertCount(2, $emails);
         $this->assertEquals('john@test.com', $emails['john@test.com']->email);
         $this->assertEquals('jane@test.com', $emails['jane@test.com']->email);